+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version 0.2 (working)
+* Support queries with multiple point values
 
 ## Version 0.1 (25-Oct-2021)
 * Initial MVP